Fix duplicate parameter component check
This did not seem to work if the the same parameter component ref was being used in multiple path's `params: [paramRef]` arrays. I believe this is because `type` variable (passed as a parameter) was `undefined` in that case (being called from `createManualParameters`, it would instead come from `zodSchema._def.openapi...`).

Being `undefined`, the `if` check would always fail if such a parameter already existed (even if it was identical).

This fixes it - it always compares with the value the parameter object would potentially be created with later on, ie. `paramType` and `paramName`.

Observation:
`const paramType = zodSchema._def?.openapi?.param?.in ?? component?.in ?? type` prefers the components type instead of the passed `type` parameter.

In case the `type` mismatches (eg. passing an existing `type: path` ref as `requestParams: { query: { someParam: pathParamRef } }`, currently, I think, would silently succeed.